move trimming attrs to trimming mod
