+++ relay-general/src/store/trimming/mod.rs
@@ -2,4 +2,4 @@ mod processor;
 mod traits;
 
 pub use processor::TrimmingProcessor;
-pub use traits::TrimmingAttrs;
+pub use traits::{BagSize, MaxChars, TrimmingAttrs};
